% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excursions.R
\name{mhcp_excursion_prob}
\alias{mhcp_excursion_prob}
\title{MHCP Multiplicative Excursion Probability}
\usage{
mhcp_excursion_prob(zeta, c = 1)
}
\arguments{
\item{zeta}{Positive MHCP parameter.}

\item{c}{Positive multiplicative excursion size.}
}
\value{
Numeric probability.
}
\description{
Computes
}
\details{
\deqn{
P(\eta_h > c\eta_{h-1})
=
\frac{2}{\pi}
\arctan\left(\frac{\zeta}{c}\right).
}
}
\examples{
mhcp_excursion_prob(0.7)
mhcp_excursion_prob(0.7, c = 5)

}
