% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{pmhcp}
\alias{pmhcp}
\title{MHCP Marginal Distribution Function}
\usage{
pmhcp(x, H, zeta)
}
\arguments{
\item{x}{Positive numeric value.}

\item{H}{Component index.}

\item{zeta}{Positive MHCP parameter.}
}
\value{
Numeric probability.
}
\description{
Computes \code{P(eta_H <= x)} for the MHCP.
}
\examples{
pmhcp(
  x = 0.05,
  H = 10,
  zeta = 0.369411
)

}
