% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{qmhcp_centered}
\alias{qmhcp_centered}
\title{Quantiles of the Centered MHCP Log Process}
\usage{
qmhcp_centered(p, H, rel_tol = 1e-08)
}
\arguments{
\item{p}{Probability in \verb{(0,1)}.}

\item{H}{Component index. Must be at least 2.}

\item{rel_tol}{Numerical tolerance.}
}
\value{
Numeric quantile.
}
\description{
Computes quantiles of
}
\details{
\deqn{
Z_H = \sum_{\ell=2}^{H} Y_\ell.
}
}
\examples{
qmhcp_centered(0.90, H = 10)

}
